+++ FytSoa.Service/DtoModel/Cms/TrendInfo.cs
@@ -5,9 +5,9 @@ namespace FytSoa.Service.DtoModel.Cms
     public class TrendInfo
     {
         public DateTime Day { get; set; }
-        public int CountTrade { get; set; }
+        public long CountTrade { get; set; }
         public long TotalTrade { get; set; }
-        public int CountRefund { get; set; }
+        public long CountRefund { get; set; }
         public long TotalRefund { get; set; }
     }
 }
